+++ app/styles/dashboard.css
@@ -4,3 +4,8 @@
 .event-name {
     padding-bottom: 1em; 
 }
+
+.calendar {
+    padding-bottom: 0.5em;
+    padding-top: 1em; 
+}
